+++ test/fr/tedoldi/lichess/game/retriever/pgn_test.clj
@@ -15,33 +15,36 @@
 (facts "PGN generation"
 
       (fact "Basic PGN"
-            (pgn->str {:event     ..event..
-                       :moves     [..m-w-1.. ..m-b-1..]
-                       :site      ..site..
-                       :timestamp 0
-                       :round     ..round..
-                       :players   {:white {:name ..white-player-name..
-                                           :elo  ..white-player-elo..}
-                                   :black {:name ..black-player-name..
-                                           :elo  ..black-player-elo..}}
-                       :result    ..result..
-                       :variant   ..variant..})
+            (pgn->str {:id      ..event..
+                       :moves   [..m-w-1.. ..m-b-1..]
+                       :speed   ..site..
+                       :date    ..date..
+                       :url     ..round..
+                       :players {:white {:name ..white-player-name..
+                                         :elo  ..white-player-elo..}
+                                 :black {:name ..black-player-name..
+                                         :elo  ..black-player-elo..}}
+                       :result  ..result..
+                       :variant ..variant..}
+                      nil)
             => (slurp (io/resource "basic.pgn")))
 
       (fact "PGN with comments"
-            (pgn->str {:event     ..event..
-                       :moves     [{:move     ..m-w-1..
-                                    :comments [..comment-m-w-1-1..
-                                               ..comment-m-w-1-2..]}
-                                   {:move     ..m-b-1..
-                                    :comments [..comment-m-b-1-1..]}]
-                       :site      ..site..
-                       :timestamp 0
-                       :round     ..round..
-                       :players   {:white {:name ..white-player-name..
-                                           :elo  ..white-player-elo..}
-                                   :black {:name ..black-player-name..
-                                           :elo  ..black-player-elo..}}
-                       :result    ..result..
-                       :variant   ..variant..})
+            (pgn->str {:id      ..event..
+                       :moves   [{:move     ..m-w-1..
+                                  :comments [..comment-m-w-1-1..
+                                             ..comment-m-w-1-2..]
+                                  :time     10.0}
+                                 {:move     ..m-b-1..
+                                  :comments [..comment-m-b-1-1..]}]
+                       :speed   ..site..
+                       :date    ..date..
+                       :url     ..round..
+                       :players {:white {:name ..white-player-name..
+                                         :elo  ..white-player-elo..}
+                                 :black {:name ..black-player-name..
+                                         :elo  ..black-player-elo..}}
+                       :result  ..result..
+                       :variant ..variant..}
+                      nil)
             => (slurp (io/resource "commented.pgn"))))