+++ CHANGELOG.md
@@ -3,9 +3,21 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+## [1.2.0] - 2016-04-11
+### Added
+- Adds game analysis aggregates as PGN comment when available
+- Allow move time display using --with-times
+- Templates are overridable (--template-pgn, --template-move-pair)
+
+## [1.1.0] - 2016-04-11
+### Changed
+- Improve computation speed
+
 ## [1.0.0] - 2016-04-09
 ### Changed
 - Initial version
 
-[Unreleased]: https://github.com/s-ted/liPGN/compare/1.0.0...HEAD
+[Unreleased]: https://github.com/s-ted/liPGN/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/s-ted/liPGN/compare/1.1.0...1.2.0
+[1.1.0]: https://github.com/s-ted/liPGN/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/s-ted/liPGN/compare/0.0.0...1.0.0